Broke up DatabaseObject loading / construction code into factories
This is a rather big, late architectural change. Most loading and
construction code - especially redundant boilerplate code - was moved
to an abstract factory and a factory template.
This allows for a nicer construction and loading syntax and more
flexibility for loading (especially the ability to build generic lazy
loaders).

Also lots of shared pointer factory methods were removed from the various
DatabaseObject subclasses and centralized in a generic construction
template.

Query building was also moved out of DatabaseObject.